third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-resolution-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <resolution> with iterationComposite
  assert_equals: expected "250dppx" but got "50dppx"
Harness: the test ran to completion.
